+++ packages/core-p2p/src/socket-server/worker.ts
@@ -5,16 +5,23 @@ import { SocketErrors } from "../enums";
 import { requestSchemas } from "../schemas";
 import { RateLimiter } from "./rate-limiter";
 
+const MINUTE_IN_MILLISECONDS = 1000 * 60;
+const HOUR_IN_MILLISECONDS = MINUTE_IN_MILLISECONDS * 60;
+
 const ajv = new Ajv();
 
 export class Worker extends SCWorker {
     private config: Record<string, any>;
+    private ipLastError: Record<string, number> = {};
     private rateLimiter: RateLimiter;
 
     public async run() {
         this.log(`Socket worker started, PID: ${process.pid}`);
 
         await this.loadConfiguration();
+
+        // purge ipLastError every hour to free up memory
+        setInterval(() => (this.ipLastError = {}), HOUR_IN_MILLISECONDS);
 
         // @ts-ignore
         this.scServer.wsServer.on("connection", (ws, req) => this.handlePayload(ws, req));
@@ -64,39 +71,49 @@ export class Worker extends SCWorker {
     }
 
     private handlePayload(ws, req) {
-        ws.on("ping", () => {
-            ws.terminate();
+        ws.prependListener("ping", () => {
+            this.setErrorForIpAndTerminate(ws, req);
         });
-        ws.on("pong", () => {
-            ws.terminate();
+        ws.prependListener("pong", () => {
+            this.setErrorForIpAndTerminate(ws, req);
         });
-        ws.on("message", message => {
-            try {
-                const InvalidMessagePayloadError: Error = this.createError(
-                    SocketErrors.InvalidMessagePayload,
-                    "The message contained an invalid payload",
-                );
-                if (message === "#2") {
-                    const timeNow: number = new Date().getTime() / 1000;
-                    if (ws._lastPingTime && timeNow - ws._lastPingTime < 1) {
-                        throw InvalidMessagePayloadError;
-                    }
-                    ws._lastPingTime = timeNow;
-                } else {
+        ws.prependListener("message", message => {
+            if (ws._disconnected) {
+                this.setErrorForIpAndTerminate(ws, req);
+            } else if (message === "#2") {
+                const timeNow: number = new Date().getTime() / 1000;
+                if (ws._lastPingTime && timeNow - ws._lastPingTime < 1) {
+                    this.setErrorForIpAndTerminate(ws, req);
+                }
+                ws._lastPingTime = timeNow;
+            } else if (message.length < 10) {
+                // except for #2 message, we should have JSON with some required properties
+                // (see below) which implies that message length should be longer than 10 chars
+                this.setErrorForIpAndTerminate(ws, req);
+            } else {
+                try {
                     const parsed = JSON.parse(message);
+                    if (parsed.event === "#disconnect") {
+                        ws._disconnected = true;
+                    }
                     if (
                         typeof parsed.event !== "string" ||
                         typeof parsed.data !== "object" ||
                         (typeof parsed.cid !== "number" &&
                             (parsed.event === "#disconnect" && typeof parsed.cid !== "undefined"))
                     ) {
-                        throw InvalidMessagePayloadError;
+                        this.setErrorForIpAndTerminate(ws, req);
                     }
+                } catch (error) {
+                    this.setErrorForIpAndTerminate(ws, req);
                 }
-            } catch (error) {
-                ws.terminate();
             }
         });
+    }
+
+    private setErrorForIpAndTerminate(ws, req): void {
+        this.ipLastError[req.socket.remoteAddress] = Date.now();
+        ws.terminate();
     }
 
     private async handleConnection(socket): Promise<void> {
@@ -117,14 +134,20 @@ export class Worker extends SCWorker {
     }
 
     private async handleHandshake(req, next): Promise<void> {
-        const isBlocked = await this.rateLimiter.isBlocked(req.socket.remoteAddress);
-        const isBlacklisted = (this.config.blacklist || []).includes(req.socket.remoteAddress);
+        const ip = req.socket.remoteAddress;
+        if (this.ipLastError[ip] && this.ipLastError[ip] > Date.now() - MINUTE_IN_MILLISECONDS) {
+            req.socket.destroy();
+            return;
+        }
+
+        const isBlocked = await this.rateLimiter.isBlocked(ip);
+        const isBlacklisted = (this.config.blacklist || []).includes(ip);
         if (isBlocked || isBlacklisted) {
             next(this.createError(SocketErrors.Forbidden, "Blocked due to rate limit or blacklisted."));
             return;
         }
 
-        const cidrRemoteAddress = cidr(`${req.socket.remoteAddress}/24`);
+        const cidrRemoteAddress = cidr(`${ip}/24`);
         const sameSubnetSockets = Object.values({ ...this.scServer.clients, ...this.scServer.pendingClients }).filter(
             client => cidr(`${client.remoteAddress}/24`) === cidrRemoteAddress,
         );